AnyTPAC/SimplePlotter: let it compile on Qt4 and Qt5 Desktop both on win32 and Ubuntu; fix compilation issues; fix datetime format

+++ AnyTPAC/SimplePlotter/mectcomm.cpp
@@ -1038,7 +1038,7 @@ bool mectComm::doRecv(struct XBlock &reply)
         tcpSocket->abort();
         return false;
     }
-   if (retval < HD_BLOCK) {
+   if (retval < (qint64)HD_BLOCK) {
         qDebug() << "doRecv() error retval=" << retval;
         tcpSocket->abort();
         return false;
